src: a bit of project structure and coverage * moved no-code files to lib/resources directory * moved version related code to specific js file * unit tests added to it ^
