+++ lib/html.js
@@ -2,7 +2,7 @@
 
 const mustache = require('mustache');
 const fs = require('fs');
-const template = fs.readFileSync('lib/template.html', 'utf8');
+const template = fs.readFileSync('lib/resources/template.html', 'utf8');
 
 function parse (object) {
   return new Promise((resolve, reject) => {
