+++ lib/resources/licenses.css
