+++ lib/whitelist.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const whitelist = require('./default-whitelist.json');
+const whitelist = require('./resources/default-whitelist.json');
 var keys = {};
 
 function check (project) {
